Get Transcript from Console, to Python
When JS code is executed to return the transcript of the video, I managed to store it in a variable in the python application (and output it). I am planning next to use implicitly_wait() instead of time.sleep() for the web driver, and will make sure we go headless mode.